requirements.txt
# Core bot
aiogram==3.26.0
# Database (async)
SQLAlchemy[asyncio]==2.0.48
aiosqlite==0.22.1
asyncpg==0.31.0
# AI providers
openai==2.26.0
google-generativeai==0.8.6
# File text extraction
PyMuPDF==1.27.2
python-docx==1.2.0
python-pptx==0.6.23
# Optional: MySQL driver (only if DATABASE_URL uses mysql+aiomysql://)
# aiomysql
# pymysql

